UPSTREAM: libpayload: drivers/usb: add a USB pre-poll hook
This adds a hook so that a payload can optionally perform USB service
functions in conjunction with regular USB port status polling. In
particular, this allows depthcharge to control the state of an
external USB mux. Some SoCs like Tiger Lake have a USB mux for Type-C
ports that must be kept in sync with the state of the port as reported
by the TCPC. This can be achieved by hooking into the poll routine to
refresh the state of the USB mux.

BUG=b:149883933
TEST=booted into recovery from Type-C flash drive on volteer

Change-Id: I427601aceb36578d8716ced649b561505a9bbc75
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: eaa219b5bba95cfdc03b9b20d1c06d84bd33c702
Original-Change-Id: Ic6c23756f64b891b3c5683cd650c605b8630b0fb
Original-Signed-off-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42072
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2232847
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>