UPSTREAM: Makefile.inc: Remove all-y with CONFIG_ARCH_xx guards
The assumption up to this point was that if the system had an x86
processor, verstage would be running on the x86 processor.  With running
verstage on the PSP, that assumption no longer holds true, so exclude
pieces of code that cause problems for verstage on the PSP.

As a generalization, remove all-y for CONFIG_ARCH_xx guarded
makefiles.

BUG=b:158124527
TEST=Build and boot on Trembyle

Change-Id: I6ad9eb470db2e60869200e241f2c0b0665596654
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: cbf6e6bdba556fac90dabd361764e4efc044241f
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Ia7dcfed699ee1c0cd5a5250431c5f05bf6d8b9c7
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42221
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2246035
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>