UPSTREAM: palkia: separate the gpio pins control for the second touch
There are two touch screen controllers on the Palkia device.
One is on the lid; another is on the base. To support
the different control path (for example: turning off the base's
touch event when we don't want to use it however still keeping
the lid's touch event), we use the different gpio pins to control
the second touch. As a result, we need to modify the devicetree
to adopt this change. With this change, we can control the
primary and secondary touch screen controller respectively.

BUG=b:149714955
TEST=lid/base touch screen works correctly

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Commit-Id: 7587526583bf7820ffcb09edcce92bbd48237382
Original-Change-Id: I1f896e334e51c78300af724cbef8d57641ae5612
Original-Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42185
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Change-Id: Iabe3eeb0b26576c1aae7bca90ab40df1db837b44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2257809
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>