UPSTREAM: ec/google/wilco: Suppress UCSI events in S0ix
If a UCSI event comes in when the EC is in S0ix mode then the kernel
driver attempts a transaction but fails and this can leave the system
in an unexpected state where the only wake source is the power button.

This change will not notify the UCSI driver if the EC is in S0ix mode
and instead keep track of the event and send it on resume.

BUG=b:157923800
TEST=tested on drallion system:
1. Put drallion system into suspend
2. Attach power supply
3. Ensure the system can wake with keypress

Change-Id: I95d9dabe24e099119d118f8dc80649980fc2e361
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3cd3cb6f33cbbd18f9acac01ca5008e2dfb76242
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: I43acb089385d9b41ac955f053e409daad67423f1
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42237
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2251628
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>