UPSTREAM: util/cbfstool: Drop IS_TOP_ALIGNED_ADDRESS() check in cbfstool_convert_fsp
This change drops the check for IS_TOP_ALIGNED_ADDRESS() before
setting offset to 0 in cbfstool_convert_fsp(). If the user provides a
baseaddress to relocate the FSP to, then the offset should be set to 0
since there is no requirement on where the file ends up in cbfs. This
allows the user to relocate the FSP to an address in lower DRAM.

BUG=b:155322763

Change-Id: If6c9635936d38772e14607ba9d86a0818f778240
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c3bb6923bdcd20f4b343ba373a7d211655d6468a
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: Ibeadbf06881f7659b2ac7d62d2152636c853fb9f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42263
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2245717
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>