UPSTREAM: amd/mandolin: unbreak SeaBIOS VBIOS support
Commit 86ba0d73f34185533e5e2d4258aa3bf3dba40ed4 added VBIOS support for
Raven2 silicon and changed the VBIOS file names to the format including
the PCI device revision number. Upstream SeaBIOS expects the file to
have only the PCI vendor and device IDs in the CBFS file name, so it
doesn't find the VBIOS any more after that patch got applied. This patch
adds the path and CBFS file name to include the Picasso VBIOS a second
time under the CBFS file name SeaBIOS expects.

This is a workaround and not a clean solution, but avoids breakage.
It's separated from the rest of the Mandolin support, so it can just be
reverted after a proper fix is implemented.

https://chromium-review.googlesource.com/2015963/ in combination with a
links file in CBFS might solve the issue for most of the cases, but it's
not sure yet if for all, so a proper fix might require more than that.

BUG=b:153675508

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Commit-Id: fe1d80cb08213a302a80714114ea3b9e632411ff
Original-Change-Id: I4d9042615965b6a2d9255c194cf23368264ffe54
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42433
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: I966e77b4d67bfe706198f383d207f88067cb5b85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2257020
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>