UPSTREAM: device: Add method to configure bus mastering based on Kconfig
The bus master bit is set at many places in coreboot's code, but the
reason for that is not quite clear. We examined not setting the
bus master bit whereever possible and tried booting without it,
which worked fine for internal PCI devices but not for PCIe. As a PCIe
device we used a Samsung M.2 NVMe SSD.

For security reasons, we would like to disable bus mastering where
possible. Depending on the device, bus mastering might get enabled
by the operating system (e.g. for iGPU) and it might be required for
some devices to work properly. However, the idea is to leave it disabled
and configure the IOMMU first before enabling it.

To have some sort of "backwards compatibility", add a method which
configures bus mastering based on an additional config option. Since
CB:42460 makes usage of this treewide, enable it by default to keep the
current behaviour for now.

Tested with Siemens/Chili, a Coffee Lake based platform.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2c02ba6069e7a04c0aa628ba93a7f21c0bdea4da
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e4a7d9f69347314e7b3110a5f68921da768cb4e8
Original-Change-Id: I876c48ea3fb4f9cf7b6a5c2dcaeda07ea36cbed3
Original-Signed-off-by: Felix Singer <felix.singer@secunet.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42459
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Subrata Banik <subrata.banik@intel.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2386723
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>