UPSTREAM: nb/intel/sandybridge/gma.c: Remove useless if condition
There's a useless check with both branches doing the same: enabling RC6
and disabling RC6p. In past, this condition would enable RC6p in IVB but
not on SNB. Then, at some point, RC6p was considered unstable and was
disabled, but the condition remained.

It's not needed so let's remove it.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0cfc5da7a8be319a000473806c9442850bfd1dda
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0df0c7e359b6e678cb6de8f946d4395259864391
Original-Change-Id: I926bb682d1b9d21185048224490b966c33204b6a
Original-Signed-off-by: Evgeny Zinoviev <me@ch1p.io>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42410
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2257632
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>