UPSTREAM: mb/pcengines/apu2: Update GPIO Reads & writes
The APU2 was using the soc/amd/common functions to do GPIO reads and
writes.  The functions that were being used are getting eliminated in
the SOC directory, but since the APU isn't using the rest of that code
(as it's not using the rest of the SOC codebase), it proved to be
problematic to use the updated functions.

The solution I've put in place here is to pull everything needed for the
GPIO reads & writes into the gpio_ftns.c & h files.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: 87f9fc8584c980dc4c73667f4c88d71d0e447a0c
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Ied39c114bdf3637977d21f56fd7db428c52e4706
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42484
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Change-Id: Icaa68187ffe85c43ad61bb25222a95c10b4f8640
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2255154
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>