UPSTREAM: Add qc_blobs repository
This patch adds a separate blobs repository for Qualcomm blobs,
analogous to the existing AMD blobs. Qualcomm's binary licenses allow
files to be redistributed and used by anyone, but they explicitly
require the user to agree to the license terms when just *downloading*
the binary (even if they're not using them to build any firmware). Some
community members do not like to have to agree to licenses for files
they're not actually using, so we are keeping these files separate from
the main blobs repository and adding an extra Kconfig to make sure the
user is aware of and must explicitly agree to this before downloading
these files.

BUG=none
BRANCH=none
TEST=none

Change-Id: I356a4f4388af6f0d7e7e3c63a9dc3f8e41a9df09
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bc1cb38ce15e059988263b04c0ea751ddf4b052d
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I247746c1b633343064c9f32ef1556000475d6c4a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42548
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2276436
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>