UPSTREAM: sb/intel/common: Add early SPI code
All Intel southbridges with SPI perform this write. Put it inside a
function in common code. Use a different name to avoid a name clash.

As it is only one statement, make it inline so that it can be defined
on the header itself. It is only called once per southbridge anyway.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie4baca6ef574d55fa6c80325c8c7078f1a5ece4e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fc7bc54e34d496a9fe2d0047eddeb9d1f48de417
Original-Change-Id: I3c284d6cffd22949d50b4c4f9846ceaef38d7cda
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42660
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2270314
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>