UPSTREAM: mb/google/zork: Move PCIE_RST0_L configuration to early GPIO table
This change moves the configuration of PCIE_RST0_L as native
function to happen in early GPIO table. This ensures that the PERST#
signal is deasserted as soon as possible when the system comes out
of sleep state in case the sleep path asserted/deasserted the PERST#
as GPIO out.

A big difference in functionality with this change is that PCIE_RST0_L
signal is now configured as part of RO, which should be fine since
all PCIe devices have a second AUX_RESET_L signal or use PCIE_RST1_L
to control the actual reset to the device.

BUG=none
BRANCH=none
TEST=none

Change-Id: I89d5fa55746b9e66a2e7e0b9ae6b6e3d6ecb430b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2ef3b2df1f562610219444d913224d0600c097a1
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: I21a9c25b5a8a6d502cdb79cbe0dbad6ef98d6d63
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42739
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2266706
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>