UPSTREAM: soc/amd/picasso: Use cbfs to locate the AMD firmware
Switch from locating the AMD firmware in the RW_A &
RW_B regions with their hardcoded locations to using CBFS to find
them.  They still need to be at the hardcoded locations so that we
can set the location inside the binary, but instead of just setting
the pointer directly to them, we now search for them with cbfs.

BUG=b:154441227
TEST=Boot & verify that binaries are located in both RW-A & RW-B

Change-Id: Ia12c2676709a99147fb065e0cc45aee236ede635
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e21698bcb7369bdbe6f1ee1c7acef80ade0af830
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I27b0593e0db7a9e6ba9b0633ac93b4d93954f002
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42831
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Eric Peers <epeers@google.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2364306
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>