UPSTREAM: acpi: Avoid freeing a device twice
The current implementation of acpi_dp_write() frees the node after it
has written it.

If the structure contains a ACPI_DP_TYPE_CHILD then a recursive call to
acpi_dp_write() frees the child and then frees it again when returning
from the call. This results in a double free.

Split the implementation into two steps, one that ones and one that
frees. This is easier to understand and fixes the bug.

Note: This likely has no effect in coreboot since it doesn't seem to
have a proper free() implementation. But it might gain one one day.

BUG=none

Change-Id: I8ac5000ba09f29e231be7645d863dcb4cfdd2fac
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8bfa51e4c485ec85c872254a4cf724a40554370f
Original-Signed-off-by: Simon Glass <sjg@chromium.org>
Original-Change-Id: Ife3917af10bc35a3c3eee38d8292f927ef15409d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42892
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2277628
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>