UPSTREAM: mb/google/volteer: Fix world-facing camera LED is always on issue
Volteer world-facing camera has a privacy LED and it is supposed
to be turned on only when the camera is being used. But the LED
is always on and this is to fix the issue.
RCAM_SNR_PWR_EN (RearCAMera_SeNsoR_PoWeR_ENable) GPIO, which
controls the world-facing camera LED, was not in the power-up
and power-down sequence definitions and this caused the issue.

BUG=b:160341981
BRANCH=none
TEST=Build and boot volteer proto 2 board. Start a camera app
and check the world-facing camera LED is only turned on only when
the camera is working.

Change-Id: Ibd13c8b4ef1e4a2d31d19b26be01827975a9108c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: feb5d10061927639cc691207d4c0f408f5578123
Original-Signed-off-by: Daniel Kang <daniel.h.kang@intel.com>
Original-Change-Id: I564690baffddfdd0f998525992643aaf16ba4b02
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42985
Original-Reviewed-by: Wonkyu Kim <wonkyu.kim@intel.com>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Daniel H Kang <daniel.h.kang@intel.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2285420
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Daniel Kang <daniel.h.kang@intel.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Daniel Kang <daniel.h.kang@intel.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>